rename: 어색한 동사 순서 변경하여 네이밍 수정 및 일괄 적용
